Update 0stor vendoring (#431)
* update to latest 0-stor

* Update tlog code for breaking change in 0-stor.

Change in metadata API.
Change timestamp from uint64 to int64

+++ tlog/time.go
@@ -10,6 +10,6 @@ func TimeNow() time.Time {
 }
 
 // TimeNowTimestamp returns current timestamp
-func TimeNowTimestamp() uint64 {
-	return uint64(TimeNow().UnixNano())
+func TimeNowTimestamp() int64 {
+	return TimeNow().UnixNano()
 }
