+++ tlog/tlogclient/reconnect_test.go
@@ -47,7 +47,7 @@ func TestReconnectFromSend(t *testing.T) {
 		}
 
 		// send
-		err := client.Send(schema.OpSet, x, int64(x), x, data)
+		err := client.Send(schema.OpSet, x, int64(x), int64(x), data)
 		require.Nil(t, err)
 	}
